Update App.js
Create function to toggle squares from App.js through id property that comes from Box.js
Set the state from previous squares, creating a new array
Then, loop through every object in the array and define current square depending on the position inside of the array
If the current square id equals the id that is given by the square, then a new variable is created spreading  values from current square, but changing the on to the opposite value of on on the current square.
Then, we push the update square into the new squares array.
If the id is NOT equal to the id given by the square, then we push the currentSquare into the new squares array.
Return new array.
When rendering the Box component, we need to make the id accessible by declaring it again with another name(id).
We add on the toggle function to toggle when clicking on the box component.