Update App.js
Another way to use the toggle function in this case, smaller.
We'll have the function toggle with the id prop, then we'll run setSquares to set the new state from the previous squares. Then we'll return a loop over the previous squares. A square will be returned if the id equals the id given. Then it will create a new array with the props from the square spread out, but the on prop will be the opposite of what it had. If the id given is not equal to the square, then, the new square will be added into the array.